refactor: move repository to crud folder
