+++ backend/api/v1/routers/orientations.py
@@ -12,10 +12,10 @@ from pydantic import BaseModel
 from api._gen.schema_gql import schema
 from api.core.emails import Member, Person, send_notebook_member_email
 from api.core.settings import settings
+from api.db.crud.orientation_info import OrientationInfoRepository
 from api.db.models.orientation_type import OrientationType
 from api.db.models.role import RoleEnum
 from api.v1.dependencies import allowed_jwt_roles
-from api.v1.repositories.orientation_info import OrientationInfoRepository
 
 router = APIRouter(
     dependencies=[
